Finished project.
Fixed missread in the homework so now the arrow in the item will delete the item and if you click on the item you will move on the second fragment.
Now the checkUpdate checks for null book error.
Implemented update method so that it works correctly.
Fixed a bug where if you go back from the fullitemfragment the app would close.